fix: replace turing by xturing
replace turing by xturing
