--- src/turing/engines/base.py
@@ -1,8 +1,0 @@
-from turing.registry import BaseParent
-
-
-class BaseEngine(BaseParent):
-    registry = {}
-
-# add_to_registry is a class method, so it's called on the class, not on an instance of the class
-# registration happens in __init__.py, so it's called when the module is imported
