--- src/turing/engines/llama_engine.py
@@ -1,28 +1,0 @@
-from pathlib import Path
-from typing import Optional, Union
-
-from turing.engines.causal import CausalEngine, CausalLoraEngine
-
-
-class LLamaEngine(CausalEngine):
-    config_name: str = "llama_engine"
-
-    def __init__(self, weights_path: Optional[Union[str, Path]] = None):
-        super().__init__("sallywww/Llama-7B", weights_path)
-
-        self.tokenizer.pad_token = self.tokenizer.eos_token
-        self.tokenizer.pad_token_id = self.tokenizer.eos_token_id
-
-    def save(self, saving_path: Union[str, Path]):
-        self.model.save_pretrained(saving_path)
-        self.tokenizer.save_pretrained(saving_path)
-
-
-class LlamaLoraEngine(CausalLoraEngine):
-    config_name: str = "llama_lora_engine"
-
-    def __init__(self, weights_path: Optional[Union[str, Path]] = None):
-        super().__init__("sallywww/Llama-7B", weights_path)
-
-        self.tokenizer.pad_token = self.tokenizer.eos_token
-        self.tokenizer.pad_token_id = self.tokenizer.eos_token_id
