+++ src/xturing/preprocessors/base.py
@@ -1,0 +1,14 @@
+from xturing.models.stable_diffusion import StableDiffusion
+from xturing.preprocessors.instruction_collator import InstructionDataCollator
+from xturing.preprocessors.text_collator import TextDataCollator
+from xturing.registry import BaseParent
+
+
+class BasePreprocessor(BaseParent):
+    registry = {}
+
+
+BasePreprocessor.add_to_registry(
+    InstructionDataCollator.config_name, InstructionDataCollator
+)
+BasePreprocessor.add_to_registry(TextDataCollator.config_name, TextDataCollator)
